Enable streaming results for MySQL resultsets.
By default, MySQL will attempt to read the entire resultset locally (which will lead to an OutOfMemoryError for large data sets).
Now, when using the MySQL JDBC driver, call the statement's enableStreamingResults method to change this behaviour.
This is done refectively so that there is no build dependency on the MySQL JDBC JAR.