Fixed usage of subqueries in UPDATE and DELETE statements.
It can be useful to reference the table being updated in expressions of an UPDATE or DELETE statement.
Previously, attempting to do this would cause the table to appear in the FROM clause of the subquery.
Now, the table being updated is treated as an 'outer' query table (as happens with correlated sub-queries).