Fixed problems generating run-time call counts.
The run-time call counts (if enabled) were being collected,
but were not being transferred correctly from the CallCountInfo
to the CallCountStats object.  As a result, no call count results
were being output.

Also, the sorting of the call counts was not based on the call
count values (as indicate) but on the function names.  This has
been fixed as well.