Evaluate 'catch' block when CAL execution cancelled.
Previously, the primCatch primitive function was not evaluating the 'catch' block when a TerminatedByClientException was caught.
The exception was simply rethrown.
This meant that the 'finally' function would not work properly when a cancel was triggered.

Now, then a TerminatedByClientException is caught by primCatch, the catch block expression is evaluated and its result discarded.
The original TerminatedByClientException is then rethrown.
This should allow finally block clean-up to work when execution is cancelled.