Fixed bottleneck with frequently-accessed CAFs and multiple threads.
In LECC, the values for a CAF are stored with the generated class in a Map keyed by execution context.
The static 'make' method is called to retrieve any existing CAF value for the context or build a new one (and cache it).
Previously, a WeakHashMap was used for this, and synchronization was needed to protect it on the static 'make' method.
This resulted in a bottleneck when multiple threads called the CAF frequently.
In particular, certain functions like concatList are written in CAF for (with no explicit arguments) and are called frequently,
resulting in other threads waiting for the global lock.

Now, a ConcurrentMap is used to store the CAF values for the execution contexts to eliminate the global locks.
In order to build a ConcurrentMap with weak references for keys, the MapMaker class from the Google Guava library is now used.
The Guava JAR has been added to the Open Quark build.

OpenQuark build 1.7.2_16