Changes for Showing Ranges and Time values.
Time values now may only include a subset of the time components in the Show output (ignoring some that are zero).
A utility function has been added to show a Time in UTC.

A Range can be shown by providing a 'show' function for the underlying values in the range.