OpenQuark build 1.7.2_47
Allow SQL statements to be cancelled when CAL execution is cancelled.

Previously, when the database is processing an expensive query or update, CAL cannot cancel execution
until the database code returns.
As a result, the database continues to process the statement and the CAL code must wait for this to finish.

Now, there is the ability to receive cancel notifications when CAL cancellation is requested.
The DataGems code makes use of this to cancel any Statement that it is executing.
This should cause the database code to terminate the statement and return control to CAL (where the rest of the CAL execution itself can be cancelled).

The cancellation notifications are disabled by default, but can be enabled using the "org.openquark.cal.enable_cancel_notification" Java runtime property.
For example:
    -Dorg.openquark.cal.enable_cancel_notification=true