Bug Fix: M4.stringTest
SCJavaDefn.buildUnboxedDirectCall() makes a call to verifyUnboxType() to
check that the unboxed input type of length() matches the return type of
max(). However, there is a type mismatch because max returns Comparable
and length expects a String and we are not smart enough to realize that we
can just downcast back to String since that was the original type anyway.

Fix by catching the exception thrown by verifyUnboxType and return null
which will cause a boxed version to be generated.