OpenQuark build 1.7.2_59
Update the executionContext used by CAL functions that are passed to Java.

Previously, when an RTComparator,  RTEquivalenceRelation, RTCalListIterator, or RTCalFunction
was created, it would bake in a reference to the ExecutionContext which created it.
This was problematic when working with these values later from a different execution context.
Some content of the context is cleared when the CAL execution 'finishes' which can cause errors
when accessing the old context in certain ways.
Also, evaluation using the old context can pick up cancellation events which were intended
for some other currently executing code on another thread.

Now, the RTExecutionContext for any currently executing code is held in a thread-local
variable.
The Java classes (such as RTComparator) now check this first for the execution context to use,
and only fall back to the original one where this isn't available.